Commit 6 - Tilpasset visninger og verifisert at all CRUD funksjonaliteter fungerer, Også lagt til to data annotations i Models

+++ Controllers/LoansController.cs
@@ -49,8 +49,8 @@ namespace BibliofilAPI.Controllers
         // GET: Loans/Create
         public IActionResult Create()
         {
-            ViewData["BookId"] = new SelectList(_context.Books, "Id", "ISBN");
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Email");
+            ViewData["BookId"] = new SelectList(_context.Books.Where(d => d.IsAvailable), "Id", "Title");
+            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id");
             return View();
         }
 
@@ -63,67 +63,29 @@ namespace BibliofilAPI.Controllers
         {
             if (ModelState.IsValid)
             {
+                var user = await _context.Users.FindAsync(loan.UserId);
+                var book = await _context.Books.FindAsync(loan.BookId);
+
+                if ((user == null) || (book == null))
+                {
+                    return NotFound();
+                }
+
+                if (book.IsAvailable == false)
+                {
+                    ModelState.AddModelError("BookId", "The book is already part of an active loan");
+                    ViewData["BookId"] = new SelectList(_context.Books.Where(d => d.IsAvailable), "Id", "Title");
+                    ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id");
+                    return View(loan);
+                }
+
+                book.IsAvailable = false;
                 _context.Add(loan);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["BookId"] = new SelectList(_context.Books, "Id", "ISBN", loan.BookId);
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Email", loan.UserId);
-            return View(loan);
-        }
-
-        // GET: Loans/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var loan = await _context.Loans.FindAsync(id);
-            if (loan == null)
-            {
-                return NotFound();
-            }
-            ViewData["BookId"] = new SelectList(_context.Books, "Id", "ISBN", loan.BookId);
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Email", loan.UserId);
-            return View(loan);
-        }
-
-        // POST: Loans/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,LoanDate,UserId,BookId")] Loan loan)
-        {
-            if (id != loan.Id)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(loan);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!LoanExists(loan.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["BookId"] = new SelectList(_context.Books, "Id", "ISBN", loan.BookId);
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Email", loan.UserId);
+            ViewData["BookId"] = new SelectList(_context.Books.Where(d => d.IsAvailable), "Id", "Title");
+            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id");
             return View(loan);
         }
 
@@ -153,9 +115,15 @@ namespace BibliofilAPI.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var loan = await _context.Loans.FindAsync(id);
+            
             if (loan != null)
             {
-                _context.Loans.Remove(loan);
+                var book = await _context.Books.FindAsync(loan.BookId);
+                if (book != null)
+                {
+                    book.IsAvailable = true;
+                    _context.Loans.Remove(loan);
+                }
             }
 
             await _context.SaveChangesAsync();
